+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## v1.0.25 (TBA)
+## v1.0.25 (2021-09-26)
 
 Now supports Phoenix 1.6.x, and `phoenix_html` 3.x.x.